feat : Adjust LoginPage
 - refactor directory : pages
