+++ results/hayoung-jeremy/src/pages/Profile/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProfilePage";
