+++ results/hayoung-jeremy/src/pages/index.ts
@@ -1,2 +1,2 @@
-export { default as LoginPage } from "./LoginPage";
-export { default as ProfilePage } from "./ProfilePage";
+export { default as LoginPage } from "./Login";
+export { default as ProfilePage } from "./Profile";
